+++ packages/terra-form-field/tests/wdio/form-field-spec.js
@@ -1,73 +1,61 @@
-describe('Form-field', () => {
-  before(() => browser.setViewportSize(Terra.viewports('medium')[0]));
-
+Terra.describeViewports('Form-field', ['medium'], () => {
   describe('Default', () => {
     before(() => browser.url('/#/raw/tests/terra-form-field/form-field/default-field'));
 
-    Terra.it.isAccessible();
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement();
   });
 
   describe('Label', () => {
     before(() => browser.url('/#/raw/tests/terra-form-field/form-field/field-label'));
 
-    Terra.it.isAccessible();
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement();
   });
 
   describe('Label Optional', () => {
     before(() => browser.url('/#/raw/tests/terra-form-field/form-field/field-label-optional'));
 
-    Terra.it.isAccessible();
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement();
   });
 
   describe('Label Required', () => {
     before(() => browser.url('/#/raw/tests/terra-form-field/form-field/field-label-required'));
 
-    Terra.it.isAccessible();
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement();
   });
 
   describe('Error Text', () => {
     before(() => browser.url('/#/raw/tests/terra-form-field/form-field/field-error-text'));
 
-    Terra.it.isAccessible();
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement();
   });
 
   describe('Help Text', () => {
     before(() => browser.url('/#/raw/tests/terra-form-field/form-field/field-help-text'));
 
-    Terra.it.isAccessible();
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement();
   });
 
   describe('Label Required Error Help Text', () => {
     before(() => browser.url('/#/raw/tests/terra-form-field/form-field/field-label-required-error-help-text'));
 
-    Terra.it.isAccessible();
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement();
   });
 
   describe('Hidden Label', () => {
     before(() => browser.url('/#/raw/tests/terra-form-field/form-field/field-hidden-label'));
 
-    Terra.it.isAccessible();
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement();
   });
 
   describe('Text Wrap', () => {
     before(() => browser.url('/#/raw/tests/terra-form-field/form-field/field-text-wrap'));
 
-    Terra.it.isAccessible();
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement();
   });
 
   describe('Inline', () => {
     before(() => browser.url('/#/raw/tests/terra-form-field/form-field/field-inline'));
 
-    Terra.it.isAccessible();
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement();
   });
 });