+++ packages/terra-doc-template/CHANGELOG.md
@@ -3,6 +3,10 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
+* Cleaned up tests
+
 
 2.11.0 - (June 28, 2019)
 ------------------